+++ testapp/urls.py
@@ -1,5 +1,6 @@
 from django.contrib import admin
 from django.urls import include, path
+from django.views.generic import TemplateView
 from rest_framework import routers
 
 from testapp.api.urls import model_router
@@ -12,4 +13,6 @@ urlpatterns = [
     path('admin/', admin.site.urls),
     # REST Viewsets
     path('api/v1/', include(router.urls)),
+    # Custom login view
+    path("other/login/", TemplateView.as_view(template_name="testapp/test_template.html"), name='other-login-view'),
 ]